+++ src/App.tsx
@@ -1,73 +1,40 @@
-import { FormEvent, useEffect, useState } from 'react'
+import { FormEvent, useState } from 'react'
 import './App.css'
 import Greeting from './components/Greeting'
 import Navbar from './components/Navbarheading'
 import Post from './components/Post'
-import { CreatePostDTO, PostDTO } from './type/dio'
-import axios from 'axios'
+import usePosts from './hooks/usePost'
 
 function App() {
-  const [posts, setPosts] = useState<PostDTO[] | null>(null)
+  const { posts, isLoading, isSubmitting, createPost } = usePosts()
   const [newTitle, setNewTitle] = useState<string>('')
   const [newBody, setNewBody] = useState<string>('')
-  const [isLoading, setIsLoading] = useState<boolean>(false)
-  const [isSubmitting, setIsSubmitting] = useState<boolean>(false)
-
-  useEffect(() => {
-    const fetchData = async () => {
-      setIsLoading(true)
-      try {
-        //const res = await fetch('https://jsonplaceholder.typicode.com/posts')
-        //const data = await res.json()
-
-        //if (!res.ok) {
-        //throw new Error('มันเอ๋อเล๋อ')
-        //}
-
-        const res = await axios.get<PostDTO[]>('https://jsonplaceholder.typicode.com/posts')
-        setPosts(res.data)
-      } catch (err) {
-        console.log(err)
-      } finally {
-        setIsLoading(false)
-      }
-    }
-
-    fetchData()
-  }, [])
 
   const handleSubmit = async (e: FormEvent) => {
     e.preventDefault()
-    setIsSubmitting(true)
+
     try {
-      const newRes: CreatePostDTO = {
-        userId: Math.floor(Math.random() * 1000),
-        title: newTitle,
-        body: newBody,
-      }
-      await axios.post<PostDTO[]>('https://jsonplaceholder.typicode.com/posts', newRes, {
-        headers: { 'Content-Type': 'application/json' },
-      })
-      console.log(newRes)
+      await createPost(newTitle, newBody)
+
+      setNewTitle('')
+      setNewBody('')
     } catch (err) {
       console.error(err)
-    } finally {
-      setIsSubmitting(false)
     }
-
-    setNewTitle('')
-    setNewBody('')
   }
+
+  if (isLoading) return <h1>Loading...</h1>
 
   return (
     <div className="App">
       <Navbar />
-      <Greeting name="Joe" isLoggedIn={true} />
+      <Greeting name="Bun" isLoggedIn={true} />
+
       <form onSubmit={handleSubmit}>
         <label>Title</label>
-        <input type="text" onChange={(e) => setNewTitle(e.target.value)} required />
+        <input type="text" value={newTitle} onChange={(e) => setNewTitle(e.target.value)} required />
         <label>Body</label>
-        <input type="text" onChange={(e) => setNewBody(e.target.value)} required />
+        <input type="text" value={newBody} onChange={(e) => setNewBody(e.target.value)} required />
 
         <button type="submit" disabled={isSubmitting}>
           {isSubmitting ? 'Submitting...' : 'Submit'}
@@ -76,8 +43,8 @@ function App() {
 
       <div className="feed-container">
         {posts &&
-          posts.map((postValue) => {
-            return <Post key={postValue.id} post={postValue} />
+          posts.map((post) => {
+            return <Post key={post.id} post={post} />
           })}
       </div>
     </div>
